fix(e2e): use beforeAll + serial pattern instead of storageState
The storageState approach (separate setup-owner project + smoke-authenticated
project reading saved cookies) fought an architectural mismatch in how
@supabase/ssr writes auth state vs. what page.context().storageState()
captures. Diagnostic logging on the prior commit revealed saved jars
came back with 0 sb-* auth cookies despite the post-login redirect
succeeding — auth must be living somewhere storageState doesn't see
(localStorage or in-memory session). Subsequent retries landed on /login
because the cookies that "worked" on attempt 1 weren't actually present.

Different architecture: one UI login at test.beforeAll level, all 4
authenticated tests share the same already-authenticated page via
test.describe.serial. No storageState involved.

Net auth-call cost per CI run:
- Before this whole PR: 5 UI logins (one per inline loginAsOwner)
- This: 2 UI logins (one for login-flow test, one for the beforeAll)

The 2-login cost is well under Supabase's ~45 sign-ins/minute limit even
when 4 PRs merge in rapid succession (4 × 2 = 8 logins/min), which was
the original flake source captured in MEMORY.md.

Reverts the prior storageState architecture:
- Deletes tests/e2e/tests/auth-ui.setup.ts
- Deletes tests/e2e/tests/smoke/authenticated-flows.smoke.spec.ts
- Restores tests/e2e/playwright.config.ts to main (drops smoke-authenticated
  project + the setup-owner project's UI variant)
- Restores .github/workflows/ci-cd.yml e2e step (drops --project=smoke-authenticated)

Combines the 4 authenticated tests back into critical-paths.smoke.spec.ts
under a test.describe.serial block with beforeAll/afterAll lifecycle.

[hudsor01]
